rpc: refactor
