+++ rpc/cmd.go
@@ -1,0 +1,18 @@
+package rpc
+
+import (
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/cosmos/cosmos-sdk/client/lcd"
+	"github.com/cosmos/cosmos-sdk/codec"
+	"github.com/spf13/cobra"
+)
+
+// ServeCmd creates a CLI command to start Cosmos REST server with web3 RPC API and
+// Cosmos rest-server endpoints
+func ServeCmd(cdc *codec.Codec) *cobra.Command {
+	cmd := lcd.ServeCommand(cdc, registerRoutes)
+	cmd.Flags().String(flagUnlockKey, "", "Select a key to unlock on the RPC server")
+	cmd.Flags().String(flagWebsocket, "8546", "websocket port to listen to")
+	cmd.Flags().StringP(flags.FlagBroadcastMode, "b", flags.BroadcastSync, "Transaction broadcasting mode (sync|async|block)")
+	return cmd
+}
