+++ rpc/namespaces/personal/personal_api.go
@@ -7,12 +7,8 @@ import (
 	"os"
 	"time"
 
-	"github.com/spf13/viper"
-
-	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/crypto/keys"
 	"github.com/cosmos/cosmos-sdk/crypto/keys/mintkey"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	"github.com/ethereum/go-ethereum/accounts"
 	"github.com/ethereum/go-ethereum/common"
@@ -21,22 +17,22 @@ import (
 
 	"github.com/cosmos/ethermint/crypto/ethsecp256k1"
 	"github.com/cosmos/ethermint/crypto/hd"
-	params "github.com/cosmos/ethermint/rpc/args"
+	"github.com/cosmos/ethermint/rpc/namespaces/eth"
 )
 
 // PersonalEthAPI is the personal_ prefixed set of APIs in the Web3 JSON-RPC spec.
 type PersonalEthAPI struct {
-	ethAPI   *PublicEthAPI
+	ethAPI   *eth.PublicEthAPI
 	keyInfos []keys.Info // all keys, both locked and unlocked. unlocked keys are stored in ethAPI.keys
 }
 
-// NewPersonalEthAPI creates an instance of the public Personal Eth API.
-func NewPersonalEthAPI(ethAPI *PublicEthAPI) *PersonalEthAPI {
+// New creates an instance of the public Personal Eth API.
+func New(ethAPI *eth.PublicEthAPI) *PersonalEthAPI {
 	api := &PersonalEthAPI{
 		ethAPI: ethAPI,
 	}
 
-	infos, err := api.getKeybaseInfo()
+	infos, err := api.ethAPI.getKeyringInfo()
 	if err != nil {
 		return api
 	}
@@ -45,34 +41,12 @@ func NewPersonalEthAPI(ethAPI *PublicEthAPI) *PersonalEthAPI {
 	return api
 }
 
-func (e *PersonalEthAPI) getKeybaseInfo() ([]keys.Info, error) {
-	e.ethAPI.keybaseLock.Lock()
-	defer e.ethAPI.keybaseLock.Unlock()
-
-	if e.ethAPI.clientCtx.Keybase == nil {
-		keybase, err := keys.NewKeyring(
-			sdk.KeyringServiceName(),
-			viper.GetString(flags.FlagKeyringBackend),
-			viper.GetString(flags.FlagHome),
-			e.ethAPI.clientCtx.Input,
-			hd.EthSecp256k1Options()...,
-		)
-		if err != nil {
-			return nil, err
-		}
-
-		e.ethAPI.clientCtx.Keybase = keybase
-	}
-
-	return e.ethAPI.clientCtx.Keybase.List()
-}
-
 // ImportRawKey armors and encrypts a given raw hex encoded ECDSA key and stores it into the key directory.
 // The name of the key will have the format "personal_<length-keys>", where <length-keys> is the total number of
 // keys stored on the keyring.
 // NOTE: The key will be both armored and encrypted using the same passphrase.
-func (e *PersonalEthAPI) ImportRawKey(privkey, password string) (common.Address, error) {
-	e.ethAPI.logger.Debug("personal_importRawKey")
+func (api *PersonalEthAPI) ImportRawKey(privkey, password string) (common.Address, error) {
+	api.ethAPI.logger.Debug("personal_importRawKey")
 	priv, err := crypto.HexToECDSA(privkey)
 	if err != nil {
 		return common.Address{}, err
@@ -83,33 +57,33 @@ func (e *PersonalEthAPI) ImportRawKey(privkey, password string) (common.Address,
 	armor := mintkey.EncryptArmorPrivKey(privKey, password, ethsecp256k1.KeyType)
 
 	// ignore error as we only care about the length of the list
-	list, _ := e.ethAPI.clientCtx.Keybase.List()
+	list, _ := api.ethAPI.clientCtx.Keybase.List()
 	privKeyName := fmt.Sprintf("personal_%d", len(list))
 
-	if err := e.ethAPI.clientCtx.Keybase.ImportPrivKey(privKeyName, armor, password); err != nil {
+	if err := api.ethAPI.clientCtx.Keybase.ImportPrivKey(privKeyName, armor, password); err != nil {
 		return common.Address{}, err
 	}
 
 	addr := common.BytesToAddress(privKey.PubKey().Address().Bytes())
 
-	info, err := e.ethAPI.clientCtx.Keybase.Get(privKeyName)
+	info, err := api.ethAPI.clientCtx.Keybase.Key(privKeyName)
 	if err != nil {
 		return common.Address{}, err
 	}
 
 	// append key and info to be able to lock and list the account
-	//e.ethAPI.keys = append(e.ethAPI.keys, privKey)
-	e.keyInfos = append(e.keyInfos, info)
-	e.ethAPI.logger.Info("key successfully imported", "name", privKeyName, "address", addr.String())
+	//api.ethAPI.keys = append(api.ethAPI.keys, privKey)
+	api.keyInfos = append(api.keyInfos, info)
+	api.ethAPI.logger.Info("key successfully imported", "name", privKeyName, "address", addr.String())
 
 	return addr, nil
 }
 
 // ListAccounts will return a list of addresses for accounts this node manages.
-func (e *PersonalEthAPI) ListAccounts() ([]common.Address, error) {
-	e.ethAPI.logger.Debug("personal_listAccounts")
+func (api *PersonalEthAPI) ListAccounts() ([]common.Address, error) {
+	api.ethAPI.logger.Debug("personal_listAccounts")
 	addrs := []common.Address{}
-	for _, info := range e.keyInfos {
+	for _, info := range api.keyInfos {
 		addressBytes := info.GetPubKey().Address().Bytes()
 		addrs = append(addrs, common.BytesToAddress(addressBytes))
 	}
@@ -119,20 +93,20 @@ func (e *PersonalEthAPI) ListAccounts() ([]common.Address, error) {
 
 // LockAccount will lock the account associated with the given address when it's unlocked.
 // It removes the key corresponding to the given address from the API's local keys.
-func (e *PersonalEthAPI) LockAccount(address common.Address) bool {
-	e.ethAPI.logger.Debug("personal_lockAccount", "address", address.String())
+func (api *PersonalEthAPI) LockAccount(address common.Address) bool {
+	api.ethAPI.logger.Debug("personal_lockAccount", "address", address.String())
 
-	for i, key := range e.ethAPI.keys {
+	for i, key := range api.ethAPI.keys {
 		if !bytes.Equal(key.PubKey().Address().Bytes(), address.Bytes()) {
 			continue
 		}
 
-		tmp := make([]ethsecp256k1.PrivKey, len(e.ethAPI.keys)-1)
-		copy(tmp[:i], e.ethAPI.keys[:i])
-		copy(tmp[i:], e.ethAPI.keys[i+1:])
-		e.ethAPI.keys = tmp
+		tmp := make([]ethsecp256k1.PrivKey, len(api.ethAPI.keys)-1)
+		copy(tmp[:i], api.ethAPI.keys[:i])
+		copy(tmp[i:], api.ethAPI.keys[i+1:])
+		api.ethAPI.keys = tmp
 
-		e.ethAPI.logger.Debug("account unlocked", "address", address.String())
+		api.ethAPI.logger.Debug("account unlocked", "address", address.String())
 		return true
 	}
 
@@ -140,9 +114,9 @@ func (e *PersonalEthAPI) LockAccount(address common.Address) bool {
 }
 
 // NewAccount will create a new account and returns the address for the new account.
-func (e *PersonalEthAPI) NewAccount(password string) (common.Address, error) {
-	e.ethAPI.logger.Debug("personal_newAccount")
-	_, err := e.getKeybaseInfo()
+func (api *PersonalEthAPI) NewAccount(password string) (common.Address, error) {
+	api.ethAPI.logger.Debug("personal_newAccount")
+	_, err := api.ethAPI.getKeyringInfo()
 	if err != nil {
 		return common.Address{}, err
 	}
@@ -153,12 +127,12 @@ func (e *PersonalEthAPI) NewAccount(password string) (common.Address, error) {
 		return common.Address{}, err
 	}
 
-	e.keyInfos = append(e.keyInfos, info)
+	api.keyInfos = append(api.keyInfos, info)
 
 	addr := common.BytesToAddress(info.GetPubKey().Address().Bytes())
-	e.ethAPI.logger.Info("Your new key was generated", "address", addr.String())
-	e.ethAPI.logger.Info("Please backup your key file!", "path", os.Getenv("HOME")+"/.ethermintcli/"+name)
-	e.ethAPI.logger.Info("Please remember your password!")
+	api.ethAPI.logger.Info("Your new key was generated", "address", addr.String())
+	api.ethAPI.logger.Info("Please backup your key file!", "path", os.Getenv("HOME")+"/.ethermintd/"+name)
+	api.ethAPI.logger.Info("Please remember your password!")
 	return addr, nil
 }
 
@@ -166,13 +140,13 @@ func (e *PersonalEthAPI) NewAccount(password string) (common.Address, error) {
 // the given password for duration seconds. If duration is nil it will use a
 // default of 300 seconds. It returns an indication if the account was unlocked.
 // It exports the private key corresponding to the given address from the keyring and stores it in the API's local keys.
-func (e *PersonalEthAPI) UnlockAccount(_ context.Context, addr common.Address, password string, _ *uint64) (bool, error) { // nolint: interfacer
-	e.ethAPI.logger.Debug("personal_unlockAccount", "address", addr.String())
+func (api *PersonalEthAPI) UnlockAccount(_ context.Context, addr common.Address, password string, _ *uint64) (bool, error) { // nolint: interfacer
+	api.ethAPI.logger.Debug("personal_unlockAccount", "address", addr.String())
 	// TODO: use duration
 
 	var keyInfo keys.Info
 
-	for _, info := range e.keyInfos {
+	for _, info := range api.keyInfos {
 		addressBytes := info.GetPubKey().Address().Bytes()
 		if bytes.Equal(addressBytes, addr[:]) {
 			keyInfo = info
@@ -184,31 +158,30 @@ func (e *PersonalEthAPI) UnlockAccount(_ context.Context, addr common.Address, p
 		return false, fmt.Errorf("cannot find key with given address %s", addr.String())
 	}
 
-	// exporting private key only works on local keys
-	if keyInfo.GetType() != keys.TypeLocal {
-		return false, fmt.Errorf("key type must be %s, got %s", keys.TypeLedger.String(), keyInfo.GetType().String())
-	}
-
-	privKey, err := e.ethAPI.clientCtx.Keybase.ExportPrivateKeyObject(keyInfo.GetName(), password)
+	privKey, err := api.ethAPI.clientCtx.Keybase.ExportPrivateKeyObject(keyInfo.GetName(), password)
 	if err != nil {
 		return false, err
 	}
 
-	emintKey, ok := privKey.(ethsecp256k1.PrivKey)
-	if !ok {
-		return false, fmt.Errorf("invalid private key type: %T", privKey)
+	if algo != ethsecp256k1.KeyType {
+		return false, fmt.Errorf("invalid key algorithm, got %s, expected %s", algo, ethsecp256k1.KeyType)
 	}
 
-	e.ethAPI.keys = append(e.ethAPI.keys, emintKey)
-	e.ethAPI.logger.Debug("account unlocked", "address", addr.String())
+	ethermintPrivKey, ok := privKey.(ethsecp256k1.PrivKey)
+	if !ok {
+		return false, fmt.Errorf("invalid private key type %T, expected %T", privKey, &ethsecp256k1.PrivKey{})
+	}
+
+	api.ethAPI.keys = append(api.ethAPI.keys, *ethermintPrivKey)
+	api.ethAPI.logger.Debug("account unlocked", "address", addr.String())
 	return true, nil
 }
 
 // SendTransaction will create a transaction from the given arguments and
 // tries to sign it with the key associated with args.To. If the given password isn't
 // able to decrypt the key it fails.
-func (e *PersonalEthAPI) SendTransaction(_ context.Context, args params.SendTxArgs, _ string) (common.Hash, error) {
-	return e.ethAPI.SendTransaction(args)
+func (api *PersonalEthAPI) SendTransaction(_ context.Context, args rpctypes.SendTxArgs, _ string) (common.Hash, error) {
+	return api.ethAPI.SendTransaction(args)
 }
 
 // Sign calculates an Ethereum ECDSA signature for:
@@ -220,10 +193,10 @@ func (e *PersonalEthAPI) SendTransaction(_ context.Context, args params.SendTxAr
 // The key used to calculate the signature is decrypted with the given password.
 //
 // https://github.com/ethereum/go-ethereum/wiki/Management-APIs#personal_sign
-func (e *PersonalEthAPI) Sign(_ context.Context, data hexutil.Bytes, addr common.Address, _ string) (hexutil.Bytes, error) {
-	e.ethAPI.logger.Debug("personal_sign", "data", data, "address", addr.String())
+func (api *PersonalEthAPI) Sign(_ context.Context, data hexutil.Bytes, addr common.Address, _ string) (hexutil.Bytes, error) {
+	api.ethAPI.logger.Debug("personal_sign", "data", data, "address", addr.String())
 
-	key, ok := checkKeyInKeyring(e.ethAPI.keys, addr)
+	key, ok := checkKeyInKeyring(api.ethAPI.keys, addr)
 	if !ok {
 		return nil, fmt.Errorf("cannot find key with address %s", addr.String())
 	}
@@ -247,8 +220,8 @@ func (e *PersonalEthAPI) Sign(_ context.Context, data hexutil.Bytes, addr common
 // the V value must be 27 or 28 for legacy reasons.
 //
 // https://github.com/ethereum/go-ethereum/wiki/Management-APIs#personal_ecRecove
-func (e *PersonalEthAPI) EcRecover(_ context.Context, data, sig hexutil.Bytes) (common.Address, error) {
-	e.ethAPI.logger.Debug("personal_ecRecover", "data", data, "sig", sig)
+func (api *PersonalEthAPI) EcRecover(_ context.Context, data, sig hexutil.Bytes) (common.Address, error) {
+	api.ethAPI.logger.Debug("personal_ecRecover", "data", data, "sig", sig)
 
 	if len(sig) != crypto.SignatureLength {
 		return common.Address{}, fmt.Errorf("signature must be %d bytes long", crypto.SignatureLength)
